+++ functions/src/index.ts
@@ -1,17 +1,6 @@
-import * as functions from 'firebase-functions';
+export { createUser } from './user.function';
+export { countUpFavorite } from './like.function';
+export { countDownFavorite } from './like.function';
 
 // Start writing Firebase Functions
 // https://firebase.google.com/docs/functions/typescript
-
-const admin = require('firebase-admin');
-admin.initializeApp();
-
-const db = admin.firestore();
-
-exports.createUser = functions.auth.user().onCreate(user => {
-  db.doc(`users/${user.uid}`).set({
-    uid: user.uid,
-    userName: user.displayName,
-    avatarURL: user.photoURL?.replace('_normal', '')
-  });
-});
