+++ functions/src/like.function.ts
@@ -1,0 +1,37 @@
+import * as functions from 'firebase-functions';
+import * as admin from 'firebase-admin';
+import { shouldEventRun, markEventTried } from './utility.function';
+
+const db = admin.firestore();
+
+export const countUpFavorite = functions.firestore
+  .document('users/{userId}/likeArticles/{articleId}')
+  .onCreate(async (snap, context) => {
+    const eventId = context.eventId;
+    return shouldEventRun(eventId).then(async (should: boolean) => {
+      if (should) {
+        await db
+          .doc(`articles/${context.params.articleId}`)
+          .update('favorite', admin.firestore.FieldValue.increment(1));
+        return markEventTried(eventId);
+      } else {
+        return;
+      }
+    });
+  });
+
+export const countDownFavorite = functions.firestore
+  .document('users/{userId}/likeArticles/{articleId}')
+  .onDelete(async (snap, context) => {
+    const eventId = context.eventId;
+    return shouldEventRun(eventId).then(async (should: boolean) => {
+      if (should) {
+        await db
+          .doc(`articles/${context.params.articleId}`)
+          .update('favorite', admin.firestore.FieldValue.increment(-1));
+        return markEventTried(eventId);
+      } else {
+        return;
+      }
+    });
+  });
